requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==4.2.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
Faker==19.3.1
httpie==3.2.2
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkg_resources==0.0.0
pluggy==1.3.0
pydotplus==2.0.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.3
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
upstox-python-sdk==2.1.0
urllib3==2.0.4
